Mise à jour des routes de création, de mise à jour et de suppression d'événements pour vérifier les autorisations dans eventRoutes.js Mise à jour des tests pour inclure l'authentification et l'autorisation dans events.test.js Mise à jour du fichier package.json dans le répertoire Tickets pour démarrer l'application en mode production Mise à jour du fichier index.js dans le répertoire de configuration pour supprimer les commentaires Mise à jour du fichier app.js dans le répertoire Tickets pour ajouter des instructions de journalisation Mise à jour du fichier index.js dans le répertoire de configuration pour supprimer les commentaires Mise à jour du fichier config/index.js dans le répertoire Tickets pour supprimer les commentaires et ajouter des instructions de journalisation

+++ Events/src/routes/eventRoutes.js
@@ -23,19 +23,30 @@ module.exports = async (app) => {
     const event = await service.getEvent(req.query);
     res.status(event.statusCode).json(event.data);
   });
-  app.post("/create", async (req, res) => {
+  app.post("/create", isAuth, async (req, res) => {
     // #swagger.tags = ['Events']
     // #swagger.description = 'Create Event.'
     // #swagger.requestBody = {required: true,content: {"application/json": {schema: {$ref: "#/components/schemas/events"}  }}}
+    const role = req.user?.role;
+
+    if (role !== "admin" || role !== "organizer")
+      return res
+        .status(403)
+        .json({ msg: "You are not authorized to create an event." });
 
     const createEvent = await service.createEvent(req.body);
     res.status(createEvent.statusCode).json(createEvent.data);
   });
-  app.put("/update", async (req, res) => {
+  app.put("/update", isAuth, async (req, res) => {
     // #swagger.tags = ['Events']
     // #swagger.description = 'Update Event'
     // #swagger.parameters['eventId'] = { description: 'Event Id' }
     // #swagger.requestBody = {required: true,content: {"application/json": {schema: {$ref: "#/components/schemas/events"}  }}}
+    const role = req.user?.role;
+    if (role !== "admin" || role !== "organizer")
+      return res
+        .status(403)
+        .json({ msg: "You are not authorized to update an event." });
 
     const eventInputs = {
       eventId: req.query?.eventId,
@@ -45,11 +56,16 @@ module.exports = async (app) => {
     res.status(updateEvent.statusCode).json(updateEvent.data);
   });
 
-  app.delete("/delete", async (req, res) => {
+  app.delete("/delete", isAuth, async (req, res) => {
     // #swagger.tags = ['Events']
     // #swagger.description = 'Delete Event by id.'
     // #swagger.parameters['eventId'] = { description: 'Event Id' }
-    console.log(req.query);
+    const role = req.user?.role;
+    if (role !== "admin" || role !== "organizer")
+      return res
+        .status(403)
+        .json({ msg: "You are not authorized to delete an event." });
+
     const { eventId } = req.query;
     const deleteEvent = await service.deleteEvent(eventId);
     res.status(deleteEvent.statusCode).json(deleteEvent.data);
